[1569] Moving Client test from Batch4 -> Batch1 as it was fixed in PR 2994
