+++ Tests/Runner/Batch4/code.js
@@ -304,118 +304,6 @@ Bridge.assembly("Bridge.ClientTest.Batch4", function ($asm, globals) {
         }
     });
 
-    Bridge.define("Bridge.ClientTest.Batch4.RefParameterTests", {
-        methods: {
-            RefTestMethod: function (r, expectBefore, write, expectAfter, writeAfter, a) {
-                Bridge.Test.NUnit.Assert.AreEqual(expectBefore, r.v);
-                r.v = write;
-                Bridge.Test.NUnit.Assert.AreEqual(write, r.v);
-                a();
-                Bridge.Test.NUnit.Assert.AreEqual(expectAfter, r.v);
-                r.v = writeAfter;
-            },
-            CanUseReferenceToThis_SPI_1569: function () {
-                // #1569
-                new Bridge.ClientTest.Batch4.RefParameterTests.S.ctor().RefThisTest();
-            }
-        }
-    });
-
-    Bridge.define("Bridge.ClientTest.Batch4.RefParameterTests.S", {
-        $kind: "struct",
-        statics: {
-            methods: {
-                M: function (s) {
-                    // Test restructure to keep assertion count correct (prevent uncaught test exception)
-                    // copy var required as C# cannot use ref vars in lambdas
-                    var copy = new Bridge.ClientTest.Batch4.RefParameterTests.S();
-                    try {
-                        copy = s.v.$clone();
-                    }
-                    catch ($e1) {
-                        $e1 = System.Exception.create($e1);
-                        copy = Bridge.getDefaultValue(Bridge.ClientTest.Batch4.RefParameterTests.S);
-                    }
-
-                    var r1 = 0;
-                    Bridge.ClientTest.Batch4.TestHelper.Safe(function () {
-                        r1 = copy.i;
-                    });
-                    Bridge.Test.NUnit.Assert.AreEqual(11, r1);
-
-                    var r2 = 0;
-                    Bridge.ClientTest.Batch4.TestHelper.Safe(function () {
-                        r2 = copy.j;
-                    });
-                    Bridge.Test.NUnit.Assert.AreEqual(12, r2);
-
-                    s.v = new Bridge.ClientTest.Batch4.RefParameterTests.S.$ctor1(42, 43);
-
-                    var copy2 = Bridge.getDefaultValue(Bridge.ClientTest.Batch4.RefParameterTests.S);
-                    try {
-                        copy2 = s.v.$clone();
-                    }
-                    catch ($e2) {
-                        $e2 = System.Exception.create($e2);
-                        copy2 = Bridge.getDefaultValue(Bridge.ClientTest.Batch4.RefParameterTests.S);
-                    }
-
-                    var r3 = 0;
-                    Bridge.ClientTest.Batch4.TestHelper.Safe(function () {
-                        r3 = copy2.i;
-                    });
-                    Bridge.Test.NUnit.Assert.AreEqual(42, r3);
-
-                    var r4 = 0;
-                    Bridge.ClientTest.Batch4.TestHelper.Safe(function () {
-                        r4 = copy2.j;
-                    });
-                    Bridge.Test.NUnit.Assert.AreEqual(43, r4);
-                },
-                getDefaultValue: function () { return new Bridge.ClientTest.Batch4.RefParameterTests.S(); }
-            }
-        },
-        fields: {
-            i: 0,
-            j: 0
-        },
-        ctors: {
-            $ctor1: function (i, j) {
-                this.$initialize();
-                this.i = i;
-                this.j = j;
-            },
-            ctor: function () {
-                this.$initialize();
-            }
-        },
-        methods: {
-            RefThisTest: function () {
-                this.i = 11;
-                this.j = 12;
-                Bridge.ClientTest.Batch4.RefParameterTests.S.M(Bridge.ref(this));
-                Bridge.Test.NUnit.Assert.AreEqual(42, this.i);
-                Bridge.Test.NUnit.Assert.AreEqual(43, this.j);
-            },
-            getHashCode: function () {
-                var h = Bridge.addHash([83, this.i, this.j]);
-                return h;
-            },
-            equals: function (o) {
-                if (!Bridge.is(o, Bridge.ClientTest.Batch4.RefParameterTests.S)) {
-                    return false;
-                }
-                return Bridge.equals(this.i, o.i) && Bridge.equals(this.j, o.j);
-            },
-            $clone: function (to) {
-                var s = to || new Bridge.ClientTest.Batch4.RefParameterTests.S();
-                s.i = this.i;
-                s.j = this.j;
-                return s;
-            }
-        }
-    });
-
     Bridge.define("Bridge.ClientTest.Batch4.Runtime.CompilerServices.RuntimeHelpersTests", {
         methods: {
             GetHashCodeCallsGetHashCodeNonVirtually_SPI_1570: function () {
